ControlDiaryEntry と ControlDiary の分割 Fixes #123
